honksquad: refactor: move PairedMarkerSystem out of nested Shared folder into Markers/
Content.Shared/RussStation had a Shared/ folder nested inside it, which is
redundant and confusing. Move PairedMarkerSystem into a meaningfully-named
Markers/ domain folder (it manages paired marker components), update its
namespace to Content.Shared.RussStation.Markers, fix the using in
SharedCarryingSystem, and delete the now-empty Shared/ folder.

Closes #867
